refactor(vlossom-mcp): move synonym-expander out of search/ into utils/
`synonym-expander.ts` is a domain-agnostic helper (query expansion and
keyword extraction) that happens to be used by the search service today.
Keeping it under `internal/search/` made it look like a search-only
concern, even though new utilities of the same shape would belong in the
same place.

Creates `src/internal/utils/` and moves the file there. Updates the only
caller and the ARCHITECTURE.md module map.

Refs: PR #413 (comment 3207905789)
