+++ packages/vlossom-mcp/src/internal/utils/synonym-expander.ts
